services/finance-service/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
celery==5.4.0
redis==5.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.2.0
